USBCore: fix overflow bug on very large strings
The Arduino code uses a uint8_t for the string max length.  Unfortunately,
it has been observed to be as 1026 bytes coming from Windows.
Previously, this would get truncated to an 8-bit value, resulting in a
3-byte string getting sent out.

Use a uint32_t instead, to avoid the overflow and display the maximum
possible string length.

+++ common/USBCore.cpp
@@ -137,9 +137,9 @@ static int USB_SendConfiguration(int maxlen) {
 }
 
 static int USB_SendStringDescriptor(const uint8_t *string_P,
-                                    uint8_t string_len) {
+                                    uint32_t string_len) {
 
-  int l = strlen((char *)string_P);
+  uint32_t l = strlen((char *)string_P);
   if (l < string_len)
     string_len = l;
